Code fixes & Int Property area impl started -Edit areas code moved to PropertyAreas.cpp -IntPropertyArea class started impl -FloatPropertyArea fixes -TileGroupProperty code -removed PropertyEditArea::clear(), no need of that
